pyproject.toml
[project]
name = "minitel-spotify"
version = "0.1.0"
description = "Affiche un titre de musique envoyé par API sur un minitel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.8.5",
    "aiosignal==1.3.1",
    "async-timeout==4.0.2",
    "attrs==23.1.0",
    "certifi==2023.7.22",
    "charset-normalizer==3.2.0",
    "fastapi[standard]>=0.121.2",
    "frozenlist==1.4.0",
    "idna==3.4",
    "multidict==6.0.4",
    "openai==0.27.8",
    "pillow==10.0.0",
    "pydantic-settings>=2.12.0",
    "pyminitel==0.2.0",
    "pyserial==3.5",
    "requests==2.31.0",
    "tqdm==4.65.0",
    "urllib3==2.0.4",
    "yarl==1.9.2",
]

[tool.uv.sources]
pyminitel = { git = "https://github.com/Zigazou/PyMinitel.git", rev = "refs/pull/9/head" }

[dependency-groups]
dev = [
]
